requirements.txt
customtkinter==5.2.0
darkdetect==0.8.0
Pillow==10.0.1
Pillow==10.0.1
